Add duplicate check

+++ app.py
@@ -28,6 +28,23 @@ def is_auth():
     return True if login else False
 
 
+def exist(db, value, kind):
+    if kind == 'Login':
+        db.execute("SELECT ID_user FROM "
+                   "todo.users WHERE Login LIKE %s;", (value,))
+    elif kind == 'Email':
+        db.execute("SELECT ID_user FROM "
+                   "todo.users WHERE Email LIKE %s;", (value,))
+    print(value, kind)
+    item = db.fetchone()
+    if item is None:
+        print('False')
+        return False
+    else:
+        print('True')
+        return True
+
+
 @app.get('/registration')
 def registration():
     return template('registration', msg='')
@@ -35,30 +52,37 @@ def registration():
 
 @app.post('/registration')
 def registration(db):
-    if request.POST.save:
-        name = request.POST.first_name.strip()
-        surname = request.POST.surname.strip()
-        email = request.POST.email.strip()
-        login = request.POST.login.strip()
-        password = request.POST.password.strip()
+    name = request.POST.first_name.strip()
+    surname = request.POST.surname.strip()
+    email = request.POST.email.strip()
+    login = request.POST.login.strip()
+    password = request.POST.password.strip()
 
-        if not validate_email(email):
-            msg = 'Неверный формат email'
-            return template('registration', msg=msg)
+    if exist(db, email, 'Email'):
+        msg = 'Данный email уже используется'
+        return template('registration', msg=msg)
 
-        if len(password) < 8:
-            msg = 'Пароль должен быть не менее 8 символов'
-            return template('registration', msg=msg)
+    if not validate_email(email):
+        msg = 'Неверный формат email'
+        return template('registration', msg=msg)
 
-        try:
-            db.execute(
-                "INSERT INTO todo.users(Name,Surname,Email,Login, Password"
-                ") VALUES (%s, %s, %s, %s, %s);", (name, surname, email,
-                                                   login, password))
-        except IntegrityError as e:
-            return template('registration', msg=str(e).split(',')[1][2:-2])
+    if exist(db, login, 'Login'):
+        msg = 'Данный логин уже используется'
+        return template('registration', msg=msg)
 
-        return redirect("/todo")
+    if len(password) < 8:
+        msg = 'Пароль должен быть не менее 8 символов'
+        return template('registration', msg=msg)
+
+    # token = auth.encode_auth_token(app, login)
+    try:
+        db.execute(
+            "INSERT INTO todo.users(Name,Surname,Email,Login, Password) VALUES (%s, %s, %s, %s, %s);",
+            (name, surname, email, login, password))
+    except IntegrityError as e:
+        return template('registration', msg=str(e).split(',')[1][2:-2])
+
+    return redirect("/todo")
 
 
 @app.get('/login')
@@ -71,7 +95,7 @@ def sign_in(db):
     login = request.POST.login.strip()
     password = request.POST.password.strip()
 
-    db.execute("SELECT Name, Surname, Email, Login, Password FROM "
+    db.execute("SELECT ID_user, Login, Password FROM "
                "todo.users WHERE Login LIKE %s;", (login,))
 
     user = db.fetchone()
@@ -79,9 +103,6 @@ def sign_in(db):
         if password == user['Password']:
             token = auth.encode_auth_token(app, user['Login'])
             print(token)
-            db.execute(
-                'UPDATE todo.users SET Token = %s  WHERE Login LIKE %s;',
-                (token, login))
             response.set_cookie("token", token, secret=key_cookie)
             return redirect("/todo")
         else:
@@ -99,21 +120,25 @@ def todo_list(db):
         db.execute(
             "SELECT ID_tasks, Task FROM todo.tasks  WHERE Status LIKE '1';")
         rows = db.fetchall()
-        if rows:
-            return template('table', rows=rows, msg='')
-        return template('table', rows=[], msg='')
+        return template('table', rows=rows, msg='')
 
 
 @app.get('/done')
 def done_list(db):
-    db.execute("SELECT ID_tasks, Task FROM todo.tasks WHERE Status LIKE '0'")
-    rows = db.fetchall()
-    return template('table', rows=rows, msg='')
+    if not is_auth():
+        return redirect('/login')
+    else:
+        db.execute(
+            "SELECT ID_tasks, Task FROM todo.tasks WHERE Status LIKE '0'")
+        rows = db.fetchall()
+        return template('table', rows=rows, msg='')
 
 
 @app.post('/new')
 def new_item(db):
-    if request.POST.save:
+    if not is_auth():
+        return redirect('/login')
+    else:
         new = request.POST.task.strip()
         db.execute("INSERT INTO todo.tasks(Task, Status) VALUES (%s,%s);",
                    (new, 1))
@@ -122,12 +147,17 @@ def new_item(db):
 
 @app.get('/new')
 def new_item():
-    return template('new_task.tpl')
+    if not is_auth():
+        return redirect('/login')
+    else:
+        return template('new_task.tpl')
 
 
 @app.post('/edit/<no:int>')
 def edit_item(no, db):
-    if request.POST.save:
+    if not is_auth():
+        return redirect('/login')
+    else:
         edit = request.POST.task.strip()
         status = request.POST.status.strip()
 
@@ -150,24 +180,31 @@ def edit_item(no, db):
 
 @app.get('/edit/<no:int>')
 def edit_item(no, db):
-    db.execute("SELECT Task FROM todo.tasks WHERE ID_tasks LIKE %s;",
-               (no,))
-    cur_data = db.fetchone()
-    return template('edit_task', old=list(cur_data.values())[0], no=no)
+    if not is_auth():
+        return redirect('/login')
+    else:
+        db.execute("SELECT Task FROM todo.tasks WHERE ID_tasks LIKE %s;",
+                   (no,))
+        cur_data = db.fetchone()
+        return template('edit_task', old=list(cur_data.values())[0], no=no)
 
 
 @app.post('/del/<no:int>')
 def del_task(no, db):
-    try:
-        db.execute(
-            "DELETE FROM todo.tasks WHERE ID_tasks LIKE %s;", (no,))
+    if not is_auth():
+        return redirect('/login')
+    else:
+        try:
+            db.execute(
+                "DELETE FROM todo.tasks WHERE ID_tasks LIKE %s;", (no,))
 
-        msg = f'Задача под номером {no} успешно удалена'
-    except:
-        msg = f'Задачу под номером {no} удалить не удалось'
-    db.execute("SELECT ID_tasks, Task FROM todo.tasks WHERE Status LIKE '0'")
-    rows = db.fetchall()
-    return template('table', rows=rows, msg=msg)
+            msg = f'Задача под номером {no} успешно удалена'
+        except:
+            msg = f'Задачу под номером {no} удалить не удалось'
+        db.execute(
+            "SELECT ID_tasks, Task FROM todo.tasks WHERE Status LIKE '0'")
+        rows = db.fetchall()
+        return template('table', rows=rows, msg=msg)
 
 
 @app.route('/static/:filename#.*#')
